#1077 hide language selector via env
