+++ public/.env.example.js
@@ -25,3 +25,4 @@ var WS_PING_TIMEOUT_PERIOD_MS = 35000
 var DISABLE_PUBLIC_VISIBILITY = false
 var ASK_FOR_TZ_UPDATE = true
 var TOOLTIP_DELAY_MS = 2000
+var HIDE_LANGUAGE_SELECTOR = false
